requirements.txt
streamlit
python-dotenv
langchain-groq
PyPDF2
langchain-core
langchain-community
